requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.0

# LangGraph / LangChain
langgraph==0.2.60
langchain-groq==0.2.4
langchain-community==0.3.14
langchain-core==0.3.29

# RAG
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# Data & Analysis
yfinance==0.2.48
pandas==2.2.3
numpy==1.26.4

# Search
tavily-python==0.5.0

# Scheduling
apscheduler==3.10.4

# HTTP
httpx==0.27.2
aiofiles==24.1.0

# Utilities
websockets==13.1
